Sync-R1/requirements.txt
accelerate
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore
albumentations
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.3.0
av==15.0.0
azure-core==1.35.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
azure-storage-queue==12.13.0
bitsandbytes
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy
cryptography==45.0.6
cycler==0.12.1
Cython==3.1.2
datasets
decord
# deepspeed==0.17.2
diffusers
dill==0.3.8
distro==1.9.0
dlib==20.0.0
docstring_parser==0.17.0
dreamsim
durationpy==0.10
e==1.4.5
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio
gradio_client==0.8.1
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio
importlib_metadata==8.7.0
importlib_resources==6.5.2
insightface
isodate==0.7.2
jaxtyping==0.3.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
latex2mathml==3.78.0
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
lpips
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib
mdurl==0.1.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
narwhals==2.0.1
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numpy>=1.24.4,<2.3.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
nvitop==1.5.1
oauthlib==3.3.1
omegaconf==2.3.0
onnx
onnxruntime
open_clip_torch==3.1.0
openai==1.97.1
opencv-python
opencv-python-headless
openpyxl==3.1.5
orjson==3.11.1
packaging==25.0
pandas==2.3.1
peft==0.17.1
pillow==10.4.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.8
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
shadowsocks==2.8.2
shapely
shellingham==1.5.4
shortuuid==1.0.13
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.47.2
stringzilla==3.12.5
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.8.0
torchaudio==2.8.0
torchmetrics
torchrun==0.2.9
torchvision==0.23.0
tqdm==4.66.4
transformers==4.55.0
triton==3.4.0
trl==0.22.1
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vertexai==1.71.1
wadler_lindig==0.1.7
wavedrom==2.0.3.post3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
xxhash==3.5.0
yarl==1.20.1
zai-sdk==0.0.3.4
zipp==3.23.0
